+++ src/Validator.php
@@ -1,0 +1,160 @@
+<?php
+
+namespace WalterLuis\Utils;
+
+/**
+ * Validator.
+ *
+ * - \p{L} or \p{Letter}: any kind of letter from any language.
+ * - \p{M} or \p{Mark}: a character intended to be combined with another character (e.g. accents, umlauts, enclosing boxes, etc.).
+ * - \p{Z} or \p{Separator}: any kind of whitespace or invisible separator.
+ * - \p{N} or \p{Number}: any kind of numeric character in any script.
+ * - \p{P} or \p{Punctuation}: any kind of punctuation character.
+ * - \p{C} or \p{Other}: invisible control characters and unused code points.
+ *
+ * @copyright Copyright © 2022 Walter Luis
+ * @license   MIT
+ * @author    Walter Luis <walterluisglez@gmail.com>
+ *
+ * @link https://www.regular-expressions.info/unicode.html
+ */
+class Validator
+{
+    /**
+     * Function verifies if given value.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function string(string $input): bool
+    {
+        /**
+         * \p{Cc} or \p{Control}: an ASCII or Latin-1 control character: 0x00–0x1F and 0x7F–0x9F.
+         */
+        return \preg_match('/^[\p{L}\p{S}\p{P}\p{Z}\p{N}\p{Cc}]+$/', $input);
+    }
+
+    /**
+     * Function verifies if given value is only text.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function text(string $input): bool
+    {
+        return \preg_match('/^[\p{L}]+$/', $input);
+    }
+
+    /**
+     * Function verifies if given value is only number including dot and comma.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function numeric(string $input): bool
+    {
+        return \preg_match('/^[\p{N}.,]+$/', $input);
+    }
+
+    /**
+     * Function verifies if given value is integer type.
+     *
+     * @param int|string $input
+     *
+     * @return bool
+     */
+    public static function integer($input): bool
+    {
+        return false !== \filter_var($input, \FILTER_VALIDATE_INT);
+    }
+
+    /**
+     * Function verifies if given value contains only words or digits.
+     *
+     * @param int|string $input
+     *
+     * @return bool
+     */
+    public static function alnum($input): bool
+    {
+        return \preg_match('/^[\p{L}\p{N}]+$/', $input);
+    }
+
+    /**
+     * Function verifies if given value can be recognized as bool.
+     *
+     * @param bool|int|string $input
+     *
+     * @return bool
+     */
+    public static function bool($input): bool
+    {
+        return null !== \filter_var($input, \FILTER_VALIDATE_BOOLEAN, \FILTER_NULL_ON_FAILURE);
+    }
+
+    /**
+     *  Function checks if given value is email.
+     *
+     * @param string $email
+     *
+     * @return bool
+     */
+    public static function email(string $email): bool
+    {
+        return false !== \filter_var($email, \FILTER_VALIDATE_EMAIL) && $email === \filter_var($email, \FILTER_SANITIZE_EMAIL);
+    }
+
+    /**
+     * Function checks if given value is url.
+     *
+     * @param string $url
+     *
+     * @return bool
+     */
+    public static function url(string $url): bool
+    {
+        return false !== \filter_var($url, \FILTER_VALIDATE_URL);
+    }
+
+    /**
+     * Function verifies if given value is compatible with default date and time format.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function dateTime(string $input): bool
+    {
+        return (false === DateTime::toISO($input, true) ? false : true);
+    }
+
+    /**
+     * Function verifies if given value is compatible with default data format.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function date(string $input): bool
+    {
+        return static::dateTime($input);
+    }
+
+    /**
+     * Function verifies if given value is compatible with default time format.
+     *
+     * @param string $input
+     *
+     * @return bool
+     */
+    public static function time(string $input): bool
+    {
+        if (Mb::mb_strlen($input) <= 5) {
+            return \preg_match('/^(2[0-3]|1[\d]|0[\d]):([0-5][\d])$/', $input);
+        }
+        return \preg_match('/^(2[0-3]|1[\d]|0[\d]):([0-5][\d]):([0-5][\d])$/', $input);
+    }
+}