+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.9] - 2022-07-14
+### Added
+- Class created with DateTime enhanced features "\WalterLuis\Utils\DateTime".
+- Class created with validator "\WalterLuis\Utils\Validator".
+
 ## [0.0.8] - 2022-07-13
 ### Added
 - Class created with Json enhanced features "\WalterLuis\Utils\Json".